requirements.txt
# AI Orchestration & Agent Framework
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.0.60

# Quantitative Data & Math
yfinance>=0.2.40
pandas>=2.2.0

# SEC API & Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Vector DB & Embeddings (RAG)
langchain-community>=0.0.10
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.0
chromadb>=0.5.0

# Utilities
python-dotenv>=1.0.0